Refactor CLI logic into SRP submodules
