+++ crates/uselesskey-cli/src/srp/mod.rs
@@ -1,0 +1,4 @@
+pub(crate) mod artifact;
+pub(crate) mod bundle;
+pub(crate) mod profiles;
+pub(crate) mod types;
